% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.sidDLNM_fit}
\alias{AIC.sidDLNM_fit}
\title{Title Calculate AIC}
\usage{
\method{AIC}{sidDLNM_fit}(object, verbose = FALSE, marginalAIC = FALSE)
}
\arguments{
\item{object}{object of class \code{sidDLNM_fit}.}
}
\description{
Title Calculate AIC
}
