% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dDLNM.R
\name{dDLNM}
\alias{dDLNM}
\title{Title fit a DLNM with fixed index parameters using dlnm package}
\usage{
dDLNM(
  formula.list,
  x.names.list,
  kE,
  kw,
  sXdat,
  Bindex,
  M,
  maxL,
  verbose,
  con.index.init
)
}
\arguments{
\item{verbose}{}
}
\description{
Title fit a DLNM with fixed index parameters using dlnm package
}
