% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sidDLNM_fit}
\alias{predict.sidDLNM_fit}
\title{Title Obtain predictions from the fitted model by sidDLNM}
\usage{
\method{predict}{sidDLNM_fit}(object, dat, CI.seed = 123, CI = 0.95, ...)
}
\arguments{
\item{object}{object of class \code{sidDLNM_fit}.}

\item{dat}{the data frame containing the variables.}

\item{CI.seed}{random seed for sampling, default \code{123}.}

\item{CI}{confidence level, default \code{0.95}.}

\item{...}{}
}
\description{
Title Obtain predictions from the fitted model by sidDLNM
}
