% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateRatio.R
\name{RateRatio.sidDLNM_fit}
\alias{RateRatio.sidDLNM_fit}
\title{Title Rate Ratio for sidDLNM_fit objects}
\usage{
\method{RateRatio}{sidDLNM_fit}(object, x0, x1, verbose = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{sidDLNM_fit}.}

\item{verbose}{whether to print messages during the process, default \code{FALSE}.}
}
\description{
Title Rate Ratio for sidDLNM_fit objects
}
